requirements.txt
APScheduler
requests
slackclient
slackbot
